Move pluralize code into lib/i18n
Templates cannot be load the en-inflectors/en-pos libraries inside
the generation sandbox used by almond-cloud
