+++ lib/i18n/american-english.js
@@ -11,6 +11,8 @@
 
 const assert = require('assert');
 const { coin } = require('../random');
+const Inflectors = require('en-inflectors').Inflectors;
+const Tag = require('en-pos').Tag;
 
 function replaceMeMy(sentence) {
     sentence = sentence.replace(/\b((?!(?:let|inform|notify|alert|send)\b)[a-zA-Z0-9]+) me\b/g, '$1 them');
@@ -303,10 +305,29 @@ const SINGLE_DEVICE_TEMPLATES = [
     ['launch $device and $command', /^(?!(?:what|when|how|who) )/],
 ];
 
+function pluralize(name) {
+    if (!name.includes(' ')) {
+        if (new Tag([name]).initial().tags[0] === 'NN')
+            return new Inflectors(name).toPlural();
+        return name;
+    } else {
+        const words = name.split(' ');
+        const tags = new Tag(words).initial().tags;
+        if (tags[tags.length - 1] !== 'NN')
+            return name;
+        else if (['VB', 'VBP', 'VBZ', 'VBD'].includes(tags[0]))
+            return name;
+        words[words.length - 1] = pluralize(words[words.length - 1]);
+        return words.join(' ');
+    }
+}
+
 module.exports = {
     postprocessSynthetic,
     detokenize,
     postprocessNLG,
+
+    pluralize,
 
     ARGUMENT_NAME_OVERRIDES,
