+++ lib/i18n/default.js
@@ -105,9 +105,16 @@ const CHANGE_SUBJECT_TEMPLATES = [];
 
 const SINGLE_DEVICE_TEMPLATES = [];
 
+function pluralize(noun) {
+    // no plural form
+    return undefined;
+}
+
 module.exports = {
     postprocessSynthetic,
     detokenize,
+
+    pluralize,
 
     ARGUMENT_NAME_OVERRIDES,
