+++ lib/i18n/italian.js
@@ -118,9 +118,16 @@ function isValidParaphrasePair(word, paraphrase) {
 const CHANGE_SUBJECT_TEMPLATES = [];
 const SINGLE_DEVICE_TEMPLATES = [];
 
+function pluralize(noun) {
+    // TODO rules are complicated and full of exceptions...
+    return undefined;
+}
+
 module.exports = {
     postprocessSynthetic,
     detokenize,
+
+    pluralize,
 
     ARGUMENT_NAME_OVERRIDES,
