+++ lib/i18n/simplified-chinese.js
@@ -77,9 +77,20 @@ function isValidParaphrasePair(word, paraphrase) {
 const CHANGE_SUBJECT_TEMPLATES = [];
 const SINGLE_DEVICE_TEMPLATES = [];
 
+function pluralize(noun) {
+    // pluralize pronouns]
+    if (/^[我你他她它]$/u.test(noun))
+        return noun + '们';
+
+    // no plural form
+    return undefined;
+}
+
 module.exports = {
     postprocessSynthetic,
     detokenize,
+
+    pluralize,
 
     ARGUMENT_NAME_OVERRIDES,
