+++ lib/i18n/traditional-chinese.js
@@ -80,9 +80,16 @@ function isValidParaphrasePair(word, paraphrase) {
 const CHANGE_SUBJECT_TEMPLATES = [];
 const SINGLE_DEVICE_TEMPLATES = [];
 
+function pluralize(noun) {
+    // TODO
+    return undefined;
+}
+
 module.exports = {
     postprocessSynthetic,
     detokenize,
+
+    pluralize,
 
     ARGUMENT_NAME_OVERRIDES,
